refactor: import api types directly in low-level modules
Untangles the upward dependency where utils.ts and offline/* were
importing types from store.ts. Now they import from generated/api-types
(or from utils.ts for the frontend-only TypeHint).
